Refs #76580, Rubocop fix for MissingSuper and Duplicate Branch
